+++ Assets/Scripts/Characters/CharacterStats.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
+// Store and Manage all Character Stats and Modifiers
 public class CharacterStats : MonoBehaviour
 {
     // Start is called before the first frame update